+++ Menu-o-Matic/Menu-o-Matic.mm
@@ -3,7 +3,7 @@
 <node TEXT="Menu-o-Matic" ID="ID_696401721" LINK="https://github.com/EdoFro/Freeplane_Menu-o-Matic" BACKGROUND_COLOR="#97c7dc">
 <font SIZE="16" BOLD="true" ITALIC="true"/>
 <hook NAME="AutomaticEdgeColor" COUNTER="16" RULE="ON_BRANCH_CREATION"/>
-<attribute_layout NAME_WIDTH="99.75 pt" VALUE_WIDTH="228.74999 pt"/>
+<attribute_layout NAME_WIDTH="104.25 pt" VALUE_WIDTH="233.24999 pt"/>
 <attribute NAME="name" VALUE="menuOMatic"/>
 <attribute NAME="version" VALUE="v0.1.5"/>
 <attribute NAME="author" VALUE="EdoFro"/>
@@ -234,8 +234,10 @@
 <node TEXT="added elapsed time message when clicking on a script based button" ID="ID_67878750"/>
 </node>
 <node TEXT="v0.1.5" ID="ID_423298702">
-<node TEXT="added TabPane lib" ID="ID_1231294444"/>
-<node TEXT="added MoMToolbar lib" ID="ID_833939518"/>
+<node TEXT="Added AutoLaunch capabillity to custom Menus" ID="ID_350307818">
+<node TEXT="They launch automatically when starting Freeplane or when opening a map that contains Custom Menus marked as &apos;AutoLaunch&apos;" ID="ID_286725258"/>
+<node TEXT="just add a &apos;launch&apos; icon to the package node to define it as an AutoLaunch menu" ID="ID_123594505"/>
+</node>
 </node>
 </node>
 <node TEXT="license" POSITION="top_or_left" ID="ID_492801974">
@@ -370,6 +372,9 @@
   </body>
 </html>
 </richcontent>
+<attribute_layout NAME_WIDTH="109.5 pt" VALUE_WIDTH="51.75 pt"/>
+<attribute NAME="runOnStartingFreeplane" VALUE="boolean"/>
+<node TEXT="&lt;?xml version=&quot;1.0&quot; encoding=&quot;UTF-8&quot;?&gt;&#xa;&lt;preferences_structure&gt;&#xa;         &lt;tabbed_pane&gt;&#xa;                  &lt;tab name=&quot;plugins&quot;&gt;&#xa;                            &lt;separator name = &quot;${name}&quot;&gt;&#xa;                                    &lt;boolean name = &quot;${name}_runOnStartingFreeplane&quot;/&gt;&#xa;                           &lt;/separator&gt;&#xa;                  &lt;/tab&gt;&#xa;         &lt;/tabbed_pane&gt;&#xa;&lt;/preferences_structure&gt;" ID="ID_1133880285" MAX_WIDTH="20 cm"/>
 </node>
 <node TEXT="default.properties" POSITION="top_or_left" ID="ID_252406354">
 <edge COLOR="#00ffff"/>
@@ -393,6 +398,8 @@
   </body>
 </html>
 </richcontent>
+<attribute_layout NAME_WIDTH="144 pt" VALUE_WIDTH="37.5 pt"/>
+<attribute NAME="${name}_runOnStartingFreeplane" VALUE="true"/>
 </node>
 <node TEXT="translations" POSITION="top_or_left" ID="ID_971883114">
 <edge COLOR="#7c0000"/>
@@ -428,20 +435,23 @@
   </body>
 </html>
 </richcontent>
-<node TEXT="en" ID="ID_1296909110">
-<attribute_layout NAME_WIDTH="195.74999 pt" VALUE_WIDTH="176.99999 pt"/>
-<attribute NAME="addons.${name}" VALUE="Menu-o-Matic"/>
-<attribute NAME="addons.${name}.launchCustomDialog" VALUE="launch menu dialog from package"/>
-<attribute NAME="addons.${name}.listCustomDialogs" VALUE="list opened menu dialogs"/>
-<attribute NAME="addons.${name}.packCustomMenu" VALUE="pack menu"/>
-<attribute NAME="addons.${name}.unpackCustomMenu" VALUE="unpack menu"/>
-<attribute NAME="addons.${name}.getMenuCommand" VALUE="menu item under mouse as new node"/>
+<node TEXT="en" ID="ID_135465452">
+<attribute_layout NAME_WIDTH="212.24999 pt" VALUE_WIDTH="411.74999 pt"/>
 <attribute NAME="addons.${name}.lauchDialogFromNodes" VALUE="launch menu dialog directly"/>
-<attribute NAME="addons.${name}.launchDialogsFromBranch" VALUE="launch all dialog packages from branch"/>
-<attribute NAME="addons.${name}.showLastUsedMenu" VALUE="show most recently used menu"/>
-<attribute NAME="addons.${name}.listCustomMenusFromMap" VALUE="list menu packages from map"/>
 <attribute NAME="addons.${name}.1launchMenuOMaticDialog" VALUE="launch Menu-o-Matic dialog"/>
+<attribute NAME="addons.${name}.launchCustomDialog" VALUE="launch menu dialog from package"/>
+<attribute NAME="addons.${name}.listCustomMenusFromMap" VALUE="list menu packages from map"/>
+<attribute NAME="addons.${name}" VALUE="Menu-o-Matic"/>
+<attribute NAME="addons.${name}.packCustomMenu" VALUE="pack menu"/>
+<attribute NAME="OptionPanel.${name}_runOnStartingFreeplane" VALUE="Open auto launch menus"/>
+<attribute NAME="addons.${name}.showLastUsedMenu" VALUE="show most recently used menu"/>
+<attribute NAME="OptionPanel.separator.${name}" VALUE="Menu-o-Matic"/>
+<attribute NAME="OptionPanel.${name}_runOnStartingFreeplane.tooltip" VALUE="Automatically opens all Custom Menus marked as AutoLaunch from all opened mindmaps on Freeplane start"/>
+<attribute NAME="addons.${name}.getMenuCommand" VALUE="menu item under mouse as new node"/>
 <attribute NAME="addons.${name}.insertMenuOMaticPackage" VALUE="Insert Menu-o-Matic package node"/>
+<attribute NAME="addons.${name}.listCustomDialogs" VALUE="list opened menu dialogs"/>
+<attribute NAME="addons.${name}.unpackCustomMenu" VALUE="unpack menu"/>
+<attribute NAME="addons.${name}.launchDialogsFromBranch" VALUE="launch all dialog packages from branch"/>
 </node>
 </node>
 <node TEXT="deinstall" POSITION="top_or_left" ID="ID_1351206087">
@@ -481,6 +491,7 @@
 <attribute NAME="delete" VALUE="${installationbase}/addons/${name}/scripts/listCustomMenusFromMap.groovy"/>
 <attribute NAME="delete" VALUE="${installationbase}/addons/${name}/scripts/launchMenuOMaticDialog.groovy"/>
 <attribute NAME="delete" VALUE="${installationbase}/addons/${name}/scripts/insertMenuOMaticPackage.groovy"/>
+<attribute NAME="delete" VALUE="${installationbase}/scripts/init/Menu-o-Matic_init.groovy"/>
 </node>
 <node TEXT="scripts" POSITION="bottom_or_right" ID="ID_1105183990">
 <edge COLOR="#007c00"/>
@@ -814,6 +825,7 @@
 </richcontent>
 <node TEXT="doc" ID="ID_108012296"/>
 <node TEXT="templates" ID="ID_1212139449"/>
+<node TEXT="scripts" ID="ID_162715608"/>
 </node>
 <node TEXT="images" POSITION="bottom_or_right" ID="ID_817685181" VGAP_QUANTITY="2 px">
 <edge COLOR="#7c7c00"/>